Move main functionality into Engine namespace
